+++ components-examples/components-example-starters-simple/src/main/java/com/zoe/framework/components/example/starters/simple/service/Example1MysqlService.java
@@ -2,8 +2,15 @@ package com.zoe.framework.components.example.starters.simple.service;
 
 import com.zoe.framework.components.example.starters.simple.entity.msyql.Example1;
 import com.zoe.framework.components.example.starters.simple.mapper.mysql.Example1Mapper;
+import com.zoe.framework.components.starter.database.annotation.DSDatabase;
 import com.zoe.framework.components.starter.database.service.impl.SuperServiceImpl;
+import com.zoe.framework.components.starter.id.service.IdService;
+import com.zoe.framework.components.util.value.data.RandomValueUtil;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.time.LocalDateTime;
+import java.util.List;
 
 /**
  * @author 蒋时华
@@ -11,6 +18,40 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class Example1MysqlService extends SuperServiceImpl<Example1Mapper, Example1> {
+
+    @Autowired
+    private IdService idService;
+
+    @DSDatabase("db0")
+    public List<Example1> addDb0() {
+        super.save(Example1.builder()
+                .id(idService.getId())
+                .name(RandomValueUtil.getChineseName())
+                .age(RandomValueUtil.getNum(1, 100))
+                .birth(LocalDateTime.now())
+                .die(RandomValueUtil.getNum(1, 100) > 50)
+                .status(RandomValueUtil.getNum(1, 100) > 50)
+                .build());
+        return super.list();
+    }
+
+    @DSDatabase("db1")
+    public List<Example1> addDb1() {
+        super.save(Example1.builder()
+                .id(idService.getId())
+                .name(RandomValueUtil.getChineseName())
+                .age(RandomValueUtil.getNum(1, 100))
+                .birth(LocalDateTime.now())
+                .die(RandomValueUtil.getNum(1, 100) > 50)
+                .status(RandomValueUtil.getNum(1, 100) > 50)
+                .build());
+        return super.list();
+    }
+
+    @DSDatabase("#dbKey")
+    public List<Example1> listByKey(String dbKey) {
+        return super.list();
+    }
 
 
 }