+++ components-starters/components-starter-annotation/src/main/java/com/zoe/framework/components/starter/annotation/configuration/AnnotationProperties.java
@@ -27,7 +27,7 @@ import java.util.List;
 @RefreshScope
 public class AnnotationProperties {
     /*** 只允许public方法 */
-    private boolean allowedPublicOnly = false;
+    private boolean allowedPublicOnly = true;
 
     @NestedConfigurationProperty
     private AccessControl accessControl = new AccessControl();