+++ components-starters/components-starter-redis/src/main/java/com/zoe/framework/components/starter/redis/configuration/commonDynamic/DynamicRedisProperties.java
@@ -61,7 +61,7 @@ public class DynamicRedisProperties {
         /**
          * 只允许public方法
          */
-        private boolean allowedPublicOnly = false;
+        private boolean allowedPublicOnly = true;
     }
 
     @Getter