+++ components-starters/components-starter-redis/src/main/java/com/zoe/framework/components/starter/redis/configuration/redissonDynamic/DynamicRedissonProperties.java
@@ -86,7 +86,7 @@ public class DynamicRedissonProperties {
         /**
          * 只允许public方法
          */
-        private boolean allowedPublicOnly = false;
+        private boolean allowedPublicOnly = true;
     }
 
     @Getter